Move all annotations related stuff under the DataAnnotations namespace
